Re-add IsBoolean check, add CreateEvaluationInterceptor
