+++ apps/backend/src/evaluation-tags/dto/create-evaluation-tag.dto.ts
@@ -5,4 +5,8 @@ export class CreateEvaluationTagDto implements ICreateEvaluationTag {
   @IsNotEmpty()
   @IsString()
   readonly value!: string;
+
+  constructor(evaluationTag: string) {
+    this.value = evaluationTag;
+  }
 }
