+++ apps/backend/src/evaluations/evaluations.controller.ts
@@ -16,7 +16,7 @@ import {FileInterceptor} from '@nestjs/platform-express';
 import {AuthzService} from '../authz/authz.service';
 import {Action} from '../casl/casl-ability.factory';
 import {JwtAuthGuard} from '../guards/jwt-auth.guard';
-import {UpdateICreateEvaluationInterceptor} from '../interceptors/convert-public-to-boolean.interceptor';
+import {CreateEvaluationInterceptor} from '../interceptors/create-evaluation-interceptor';
 import {PublicBooleanPipe} from '../pipes/public-boolean.pipe';
 import {User} from '../users/user.model';
 import {CreateEvaluationDto} from './dto/create-evaluation.dto';
@@ -58,7 +58,7 @@ export class EvaluationsController {
   }
 
   @Post()
-  @UseInterceptors(UpdateICreateEvaluationInterceptor, FileInterceptor('data'))
+  @UseInterceptors(FileInterceptor('data'), CreateEvaluationInterceptor)
   async create(
     @Body(new PublicBooleanPipe())
     createEvaluationDto: CreateEvaluationDto,
